+++ tests/tests_models.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 from django.test import TestCase
-from djangocms_video.models import Video
+from djangocms_video.models import VideoPlayer
 
 
 EXAMPLE_VIDEO = 'http://www.youtube.com/watch?v=-iJ7bs4mTUY'
@@ -9,13 +9,12 @@ EXAMPLE_VIDEO = 'http://www.youtube.com/watch?v=-iJ7bs4mTUY'
 class VideoTestCase(TestCase):
 
     def setUp(self):
-        Video.objects.create(
-            movie_url=EXAMPLE_VIDEO,
-            width=200,
-            height=200,
+        VideoPlayer.objects.create(
+            template='default',
+            embed_link=EXAMPLE_VIDEO,
         )
 
     def test_video_instance(self):
         """Video player instance has been created"""
-        video = Video.objects.get(movie_url=EXAMPLE_VIDEO)
-        self.assertEqual(video.movie_url, EXAMPLE_VIDEO)
+        video = VideoPlayer.objects.get(embed_link=EXAMPLE_VIDEO)
+        self.assertEqual(video.embed_link, EXAMPLE_VIDEO)